refactor: notification 권한 요청 시점 모래 사장 첫 진입 시점으로 수정
